backend/requirements.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.2
certifi==2023.11.17
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
filelock==3.13.1
Flask==3.0.3
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pipenv==2023.11.15
platformdirs==4.1.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.0.3
setuptools==69.0.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.4
